re-organize evolutionary base structure
