+++ GAlgorithm/__init__.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 
-from .bases import AbstractIndividual
-from .bases import Population
+from .population import AbstractIndividual
+from .population import Population
 from .geneticalgorithm import GeneticAlgorithm
 
 from .saver import load, dump
